+++ CHANGELOG.md
@@ -60,3 +60,6 @@
 
 ## Unreleased
 - Refactor forgot/reset password screens to use components/ui primitives.
+
+## Unreleased
+- Start migrating admin list filters to components/ui Input.